Bind to `Kqueue::removeDetached` in `ntco::Kqueue` constructor
The `Kqueue` constructor keeps a function object `d_detachFunctor` that
removes an entry from the `Kqueue` and announce its detachment.
Initializing this function object takes one of two codepaths, selected
at compile time:

  1. If the compiler supports C++11 lambdas, we assign to
     `d_detachFunctor` a lambda object that calls
     `this->removeDetached`.

  2. Otherwise, the compiler does not support C++11 lambdas, so we
     assign to `d_detachFunctor` the result of binding `this` to
     `Epoll::removeDetached`.

The second code path will fail to compile on a compiler that only
supports C++03, as `Epoll` is only defined in `ntco_epoll.cpp`, and as
such is not available to the implementation of `ntco_kqueue`.
Furthermore, because in the first case, we are using a lambda to bind
`Kqueue::removeDetached`, it is likely that this behavior was
unintentional anyway.

This patch changes the bind call from `Epoll::removeDetached` to
`Kqueue::removeDetached`, which fixes the compiler error on C++03-only
compilers and provides parity between the two compiler paths.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>